Restructure the integration test schema files
