+++ tests/integration_testnet/api/rpc/votes.test.js
@@ -24,7 +24,7 @@ import {
 	invalidParamsSchema,
 	jsonRpcEnvelopeSchema,
 } from './schemas/generics.schema';
-import { goodRequestSchema } from '../../helpers/schemas';
+const { goodRequestSchema } = require('../../helpers/schemas/general');
 
 const wsRpcUrl = `${config.SERVICE_ENDPOINT}/rpc-v1`;
